+++ zest/releaser/release.py
@@ -155,6 +155,9 @@ class Releaser(baserelease.Basereleaser):
                 self.data["tagworkingdir"],
             )
             builder.build("wheel", "./dist/")
+        if not self.zest_releaser_config.upload_pypi():
+            logger.info("Upload to PyPI was disabled in the configuration.")
+            return
         if not self.pypiconfig.is_pypi_configured():
             logger.error(
                 "You must have a properly configured %s file in "